+++ lib/feature/home_page/view/ui/home_page.dart
@@ -28,7 +28,8 @@ class HomePage extends StatelessWidget {
                   children: [
                     const CircleAvatar(
                       backgroundImage: AssetImage(
-                          'assets/images/login_section/png/profile_avatar.png'),
+                        'assets/images/login_section/png/profile_avatar.png',
+                      ),
                     ),
                     const SizedBox(width: AltaSpacing.space12),
                     Column(
