+++ lib/utils/constant/alta_spacing.dart
@@ -1,11 +1,8 @@
 class AltaSpacing {
   static const space4 = 4.0;
-  static const space5 = 5.0;
-  static const space6 = 6.0;
   static const space8 = 8.0;
   static const space12 = 12.0;
   static const space16 = 16.0;
-  static const space18 = 18.0;
   static const space20 = 20.0;
   static const space24 = 24.0;
   static const space28 = 28.0;
