+++ lib/utils/widgets/alta_text.dart
@@ -62,7 +62,7 @@ class AltaText extends StatelessWidget {
         case AltaTextStyle.titleH3:
           return TextStyle(
             color: color,
-            fontWeight: FontWeight.w400,
+            fontWeight: FontWeight.normal,
             fontSize: 16,
           );
         case AltaTextStyle.bodyH1:
