+++ pubspec.yaml
@@ -37,6 +37,7 @@ dependencies:
   cupertino_icons: ^1.0.2
   flutter_svg: ^2.0.5
   firebase_core: ^2.9.0
+  url_launcher: ^6.1.10
 
 dev_dependencies:
   flutter_test:
